% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbConnection.r
\name{buildDBUrlGeneric}
\alias{buildDBUrlGeneric}
\title{buildDBUrlGeneric}
\usage{
buildDBUrlGeneric(db)
}
\arguments{
\item{db}{An \code{OrientDB} object}
}
\value{
A string with a place holder for the query type
}
\description{
Builds URL for OrientDB API
}
\details{
Builds URL for OrientDB API
}
\examples{
db <- dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
             username='admin', password='admin', port='2480')
buildDBUrlGeneric(db)
}
\author{
Jared P. Lander
}

