% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.r
\name{buildQuery}
\alias{buildQuery}
\title{buildQuery}
\usage{
buildQuery(db, query, limit = -1)
}
\arguments{
\item{db}{An OrientDB object created by \code{\link{dbInfo}}}

\item{query}{The SQL+ query to run}

\item{limit}{The limit for the number of records; defualt is -1 for all records}
}
\value{
A URL for the OrientDB REST API
}
\description{
Builds a URL string with a query built in
}
\details{
Takes an OrientDB object, builds it's connection string and adds the desired query after URL encoding it
}
\examples{
db <- dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
             username='admin', password='admin', port='2480')
buildQuery(db, 'SELECT * FROM V')
}
\author{
Jared P. Lander
}

