% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbConnection.r
\name{dbInfo}
\alias{dbInfo}
\title{dbInfo}
\usage{
dbInfo(host = "localhost", database = "GratefulDeadConcerts",
  username = "admin", password = "admin", port = "2480")
}
\arguments{
\item{host}{Hostname either as an IP address or a URL without the 'http'}

\item{database}{Name of the database}

\item{username}{username}

\item{password}{password}

\item{port}{Port number for connecting over http}
}
\value{
An object of class \code{OrientDB} to be used as a connection string
}
\description{
Builds connection information for a database
}
\details{
Essentially builds the URL for a GET call
}
\examples{
dbInfo(host='127.0.0.1',
     database='GratefulDeadConcerts',
     username='admin', password='admin', port='2480')
}
\author{
Jared P. Lander
}

