% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.r
\name{format.logical}
\alias{format.logical}
\title{format.logical}
\usage{
\method{format}{logical}(x, ...)
}
\arguments{
\item{x}{An object to be formatted}

\item{\dots}{Further arguments}
}
\value{
Simply retursn \code{x}
}
\description{
Formating function to handle \code{NA}s
}
\details{
format fails when given an \code{NA} which is a \code{logical} so this function just returns the \code{NA}
}
\examples{
format(TRUE)
format(FALSE)
format(NA)

format(c(TRUE, FALSE, NA))
}
\author{
Jared P. Lander
}

