% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert.r
\name{insert}
\alias{insert}
\title{insert}
\usage{
insert(db, object, type = c("document", "graph"))
}
\arguments{
\item{db}{An OrientDB Object generated by \code{\link{dbInfo}}}

\item{object}{A named list}

\item{type}{Type of database to be written, can only be 'document' (default) or 'graph'}
}
\value{
A response object indicating success or failure
}
\description{
Inserts records into the database
}
\details{
Inserts records into the database, one at a time
}
\examples{
\dontrun{
db <- dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
             username='admin', password='admin', port='2480')
# note, below where the list name is 'class', there should be an at-symbol in front of 'class'
# but the R engine will not allow this
insert(db, object=list('class'='V', type='song', song_type='cover', name='Super Cool Song'))
}
}
\author{
Jared P. Lander
}

