% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbConnection.r
\name{print.OrientDB}
\alias{print.OrientDB}
\title{print.OrientDB}
\usage{
\method{print}{OrientDB}(x, ...)
}
\arguments{
\item{x}{\code{OrientDB} object}

\item{\dots}{Further arguments}
}
\value{
Prints the connection string
}
\description{
Print method for OrientDB connection string
}
\details{
S3 method for printing conenction string info
}
\examples{
dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
         username='admin', password='admin', port='2480')
print(dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
             username='admin', password='admin', port='2480'))
}
\author{
Jared P. Lander
}

