% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.r
\name{query}
\alias{query}
\title{query}
\usage{
query(db, query, limit = -1)
}
\arguments{
\item{db}{An OrientDB Object generated by \code{\link{dbInfo}}}

\item{query}{The query to be executed}

\item{limit}{The maximum number of records to be returned.  BY default it is set to -1 to return all records}
}
\value{
A response object holding the results
}
\description{
Query an OrientDB database
}
\details{
Runs the supplied query againast the database
}
\examples{
\dontrun{
db <- dbInfo(host='127.0.0.1', database='GratefulDeadConcerts',
             username='admin', password='admin', port='2480')
query(db, 'SELECT * FROM V')
}
}
\author{
Jared P. Lander
}

